Rename Java package in presto-geospatial
Renamed com.facebook.presto.geospatial package in presto-geospatial module to xxx.plugin.geospatial to avoid having it split between presto-geospatial and presto-geospatial-toolkit modules.
