+++ presto-geospatial/src/test/java/com/facebook/presto/plugin/geospatial/BenchmarkGeometryToBingTiles.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.geospatial;
+package com.facebook.presto.plugin.geospatial;
 
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
